BE delete gallery

+++ resources/views/backend/gallerys/index.blade.php
@@ -68,20 +68,10 @@
                                                 <a href="" class="btn btn-warning">
                                                     <i class="fa fa-pencil text-white" aria-hidden="true"></i>
                                                 </a>
-                                                <a
-                                                href="{{route('gallery.destroy', ['id'=>$gallery->id])}}"
-                                                class="btn btn-danger btn-delete"
-                                                onclick=""
-                                                >
+                                                <a href="{{route('gallery.destroy', ['id'=>$gallery->id])}}" class="btn btn-danger btn-delete" onclick="">
                                                     <i class="fa fa-trash-o" aria-hidden="true"></i>
                                                 </a>
-                                                {!!Form::open([
-                                                    'method' => 'DELETE',
-                                                    'route' => ['gallery.destroy',$gallery->id],
-                                                    'onsubmit' => 'return confirmDelete()',
-                                                    'id' => 'form-delete'
-                                                ])!!}
-                                                {!! Form::close() !!}
+                                                {!!Form::open([ 'method' => 'DELETE', 'route' => ['gallery.destroy',$gallery->id], 'onsubmit' => 'return confirmDelete()', 'id' => 'form-delete' ])!!} {!! Form::close() !!}
                                             </td>
                                         </tr>
                                         @include('backend.gallerys.detail') @endforeach
@@ -99,6 +89,7 @@
     </div>
 </main>
 @endsection
+
 @push('script')
 <script type="text/javascript">
     $('.btn-delete').on('click', function(e) {
@@ -106,15 +97,14 @@
         $('#form-delete').submit();
     });
 
-    function confirmDelete()
-    {
+    function confirmDelete() {
         var x = confirm("Are you sure you want to delete?");
         if (x) {
             return true;
-        }
-        else {
+        } else {
             return false;
         }
     }
 </script>
 @endpush
+
